Extract 2 files for ImageSlice and ImageView
